pyproject.toml
[tool.poetry]
name = "testtaskbotai"
version = "0.1.0"
description = ""
authors = ["Victor Polyakov <polyakov.victor21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.10.0"
python-dotenv = "^1.0.1"
aiogram = "2.25"
aiofiles = "^23.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
